♻️ refactor(ar-generator-service): Refactor LLM invocation and embedding logic
